+++ test/config_test.rb
@@ -155,7 +155,7 @@ class ConfigTest < Minitest::Test
     exn = assert_raises Runners::Config::BrokenYAML do
       Runners::Config.new(path: Pathname(FILE_NAME), raw_content: "@").content
     end
-    assert_equal "Your `sider.yml` is broken at line 1 and column 1. Please fix and retry.", exn.message
+    assert_equal "`sider.yml` is broken at line 1 and column 1", exn.message
   end
 
   def test_path_name
@@ -286,5 +286,15 @@ class ConfigTest < Minitest::Test
   def test_invalid
     assert Runners::Config.new(path: Pathname(FILE_NAME), raw_content: "`").invalid?
     refute Runners::Config.new(path: Pathname(FILE_NAME), raw_content: "").invalid?
+  end
+
+  def test_parse
+    assert_nil Runners::Config.new(path: Pathname(FILE_NAME), raw_content: nil).parse
+    assert_nil Runners::Config.new(path: Pathname(FILE_NAME), raw_content: "").parse
+    assert_equal [1, "foo", true], Runners::Config.new(path: Pathname(FILE_NAME), raw_content: "[1, foo, true]").parse
+
+    assert_raises Runners::Config::BrokenYAML do
+      Runners::Config.new(path: Pathname(FILE_NAME), raw_content: "`").parse
+    end
   end
 end